requirements.txt
osxmmkeys>=0.0.2
python-mpd2>=0.5
argparse
